Refactor nodes into own folder
