+++ src/appServiceExplorer.ts
@@ -5,8 +5,8 @@
 
 import { TreeDataProvider, TreeItem, EventEmitter, Event } from 'vscode';
 import { AzureAccountWrapper } from './azureAccountWrapper';
-import { AppServiceNode, SubscriptionNode, NotSignedInNode, LoadingNode } from './appServiceNodes';
-import { NodeBase } from './nodeBase';
+import { AppServiceNode, SubscriptionNode, NotSignedInNode, LoadingNode } from './nodes/appServiceNodes';
+import { NodeBase } from './nodes/nodeBase';
 
 export class AppServiceDataProvider implements TreeDataProvider<NodeBase> {
     private _onDidChangeTreeData: EventEmitter<NodeBase> = new EventEmitter<NodeBase>();
